analysis package XLS: Add option ntdll-protect=0
There is an issue regarding the "ntdll write protection preventing AppV hooking
in Office 2016 32bit #21" in the capemon repository.
https://github.com/kevoreilly/capemon/issues/21

It states "Starting Winword and Excel 2016 32 bit with capemon loaded on recent
Windows 10 quickly ends in an error message The operating system is not
presently configured to run this application".

"... in the unhooked Winword.exe, disassembly of ntdll exports contain hooks
redirecting into module appvisvsubsystems32. In the monitored process, they do
not"